rasterising

+++ rasteriser/Source/TestModelH.h
@@ -10,7 +10,7 @@ typedef struct Pixel
 {
     int x;
     int y;
-    int zinv;
+    float zinv;
 } Pixel;
 
 // Used to describe a triangular surface:
